async_nats::client

Struct Client

pub struct Client { /* private fields */ }

Client is a Cloneable handle to NATS connection. Client should not be created directly. Instead, one of two methods can be used: crate::connect and crate::ConnectOptions::connect

Implementations

impl Client

pub fn timeout(&self) -> Option<Duration>

Returns the default timeout for requests set when creating the client.

Examples

let client = async_nats::connect("demo.nats.io").await?;
println!("default request timeout: {:?}", client.timeout());

pub fn server_info(&self) -> ServerInfo

Returns last received info from the server.

Examples

let client = async_nats::connect("demo.nats.io").await?;
println!("info: {:?}", client.server_info());

pub fn is_server_compatible(&self, major: i64, minor: i64, patch: i64) -> bool

Returns true if the server version is compatible with the version components.

This has to be used with caution, as it is not guaranteed that the server that client is connected to is the same version that the one that is a JetStream meta/stream/consumer leader, especially across leafnodes.

Examples

let client = async_nats::connect("demo.nats.io").await?;
assert!(client.is_server_compatible(2, 8, 4));

pub async fn publish<S: ToSubject>( &self, subject: S, payload: Bytes, ) -> Result<(), PublishError>

Publish a Message to a given subject.

Examples

let client = async_nats::connect("demo.nats.io").await?;
client.publish("events.data", "payload".into()).await?;

pub async fn publish_with_headers<S: ToSubject>( &self, subject: S, headers: HeaderMap, payload: Bytes, ) -> Result<(), PublishError>

Publish a Message with headers to a given subject.

Examples

use std::str::FromStr;
let client = async_nats::connect("demo.nats.io").await?;
let mut headers = async_nats::HeaderMap::new();
headers.insert(
    "X-Header",
    async_nats::HeaderValue::from_str("Value").unwrap(),
);
client
    .publish_with_headers("events.data", headers, "payload".into())
    .await?;

pub async fn publish_with_reply<S: ToSubject, R: ToSubject>( &self, subject: S, reply: R, payload: Bytes, ) -> Result<(), PublishError>

Publish a Message to a given subject, with specified response subject to which the subscriber can respond. This method does not await for the response.

Examples

let client = async_nats::connect("demo.nats.io").await?;
client
    .publish_with_reply("events.data", "reply_subject", "payload".into())
    .await?;

pub async fn publish_with_reply_and_headers<S: ToSubject, R: ToSubject>( &self, subject: S, reply: R, headers: HeaderMap, payload: Bytes, ) -> Result<(), PublishError>

Publish a Message to a given subject with headers and specified response subject to which the subscriber can respond. This method does not await for the response.

Examples

use std::str::FromStr;
let client = async_nats::connect("demo.nats.io").await?;
let mut headers = async_nats::HeaderMap::new();
client
    .publish_with_reply_and_headers("events.data", "reply_subject", headers, "payload".into())
    .await?;

pub async fn request<S: ToSubject>( &self, subject: S, payload: Bytes, ) -> Result<Message, RequestError>

Sends the request with headers.

Examples

let client = async_nats::connect("demo.nats.io").await?;
let response = client.request("service", "data".into()).await?;

pub async fn request_with_headers<S: ToSubject>( &self, subject: S, headers: HeaderMap, payload: Bytes, ) -> Result<Message, RequestError>

Sends the request with headers.

Examples

let client = async_nats::connect("demo.nats.io").await?;
let mut headers = async_nats::HeaderMap::new();
headers.insert("Key", "Value");
let response = client
    .request_with_headers("service", headers, "data".into())
    .await?;

pub async fn send_request<S: ToSubject>( &self, subject: S, request: Request, ) -> Result<Message, RequestError>

Sends the request created by the Request.

Examples

let client = async_nats::connect("demo.nats.io").await?;
let request = async_nats::Request::new().payload("data".into());
let response = client.send_request("service", request).await?;

pub fn new_inbox(&self) -> String

Create a new globally unique inbox which can be used for replies.

Examples

let reply = nc.new_inbox();
let rsub = nc.subscribe(reply).await?;

pub async fn subscribe<S: ToSubject>( &self, subject: S, ) -> Result<Subscriber, SubscribeError>

Subscribes to a subject to receive messages.

Examples

use futures_util::StreamExt;
let client = async_nats::connect("demo.nats.io").await?;
let mut subscription = client.subscribe("events.>").await?;
while let Some(message) = subscription.next().await {
    println!("received message: {:?}", message);
}

pub async fn queue_subscribe<S: ToSubject>( &self, subject: S, queue_group: String, ) -> Result<Subscriber, SubscribeError>

Subscribes to a subject with a queue group to receive messages.

Examples

use futures_util::StreamExt;
let client = async_nats::connect("demo.nats.io").await?;
let mut subscription = client.queue_subscribe("events.>", "queue".into()).await?;
while let Some(message) = subscription.next().await {
    println!("received message: {:?}", message);
}

pub async fn flush(&self) -> Result<(), FlushError>

Flushes the internal buffer ensuring that all messages are sent.

Examples

let client = async_nats::connect("demo.nats.io").await?;
client.flush().await?;

pub async fn drain(&self) -> Result<(), DrainError>

Drains all subscriptions, stops any new messages from being published, and flushes any remaining messages, then closes the connection. Once completed, any streams associated with the connection and its Clients will be closed, and further crate::Client commands will fail.

Examples

use futures_util::StreamExt;
let client = async_nats::connect("demo.nats.io").await?;
let mut subscription = client.subscribe("events.>").await?;

client.drain().await?;

assert!(subscription.next().await.is_none());
client
    .subscribe("events.>")
    .await
    .expect_err("Expected further commands to fail");

pub fn connection_state(&self) -> State

Returns the current state of the connection.

Examples

let client = async_nats::connect("demo.nats.io").await?;
println!("connection state: {}", client.connection_state());

pub async fn force_reconnect(&self) -> Result<(), ReconnectError>

Forces the client to reconnect. Keep in mind that client will reconnect automatically if the connection is lost and this method does not have to be used in normal circumstances. However, if you want to force the client to reconnect, for example to re-trigger the auth-callback, or manually rebalance connections, this method can be useful. This method does not wait for connection to be re-established.

Examples

let client = async_nats::connect("demo.nats.io").await?;
client.force_reconnect().await?;

pub fn statistics(&self) -> Arc<Statistics>

Returns struct representing statistics of the whole lifecycle of the client. This includes number of bytes sent/received, number of messages sent/received, and number of times the connection was established. As this returns Arc with [AtomicU64] fields, it can be safely reused and shared across threads.

Examples

let client = async_nats::connect("demo.nats.io").await?;
let statistics = client.statistics();
println!("client statistics: {:#?}", statistics);