async_nats

Module header

NATS Message headers, modeled loosely after the http::header crate.

Structs

GetAll

HeaderMap

A struct for handling NATS headers. Has a similar API to http::header, but properly serializes and deserializes according to NATS requirements.

HeaderName

Defines a NATS header field name

HeaderValue

Represents NATS header field value.

ParseHeaderNameError

ParseHeaderValueError

Constants

NATS_EXPECTED_LAST_MESSAGE_ID

The expected last message ID within the stream.

NATS_EXPECTED_LAST_SEQUENCE

The expected last sequence number within the stream.

NATS_EXPECTED_LAST_SUBJECT_SEQUENCE

The expected last sequence number of the subject.

NATS_EXPECTED_STREAM

The expected stream the message should be part of.

NATS_LAST_CONSUMER

The last known consumer that processed the message.

NATS_LAST_SEQUENCE

The last known sequence number of the message.

NATS_LAST_STREAM

The last known stream the message was part of.

NATS_MARKER_REASON

Reason why the delete marked on a stream with enabled markers was put.

NATS_MESSAGE_ID

A unique identifier for the message.

NATS_MESSAGE_TTL

Sets the TTL for a single message.

NATS_SEQUENCE

The sequence number of the message within the stream.

NATS_STREAM

The name of the stream the message belongs to.

NATS_SUBJECT

The subject of the message, used for routing and filtering messages.

NATS_TIME_STAMP

The timestamp of when the message was sent.

Traits

IntoHeaderName

IntoHeaderValue