async_nats::header

Struct HeaderMap

pub struct HeaderMap { /* private fields */ }

A struct for handling NATS headers. Has a similar API to http::header, but properly serializes and deserializes according to NATS requirements.

Examples

let client = async_nats::connect("demo.nats.io").await?;
let mut headers = async_nats::HeaderMap::new();
headers.insert("Key", "Value");
client
    .publish_with_headers("subject", headers, "payload".into())
    .await?;

Implementations

impl HeaderMap

pub fn iter(&self) -> Iter<'_, HeaderName, Vec<HeaderValue>>

impl HeaderMap

pub fn new() -> Self

Create an empty HeaderMap.

Examples

let map = HeaderMap::new();

assert!(map.is_empty());

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

let mut map = HeaderMap::new();

assert!(map.is_empty());

map.insert(NATS_SUBJECT, "FOO.BAR");

assert!(!map.is_empty());

pub fn len(&self) -> usize

impl HeaderMap

pub fn insert<K: IntoHeaderName, V: IntoHeaderValue>( &mut self, name: K, value: V, )

Inserts a new value to a HeaderMap.

Examples

use async_nats::HeaderMap;

let mut headers = HeaderMap::new();
headers.insert("Key", "Value");

pub fn append<K: IntoHeaderName, V: IntoHeaderValue>( &mut self, name: K, value: V, )

Appends a new value to the list of values to a given key. If the key did not exist, it will be inserted with provided value.

Examples

use async_nats::HeaderMap;

let mut headers = HeaderMap::new();
headers.append("Key", "Value");
headers.append("Key", "Another");

pub fn get<K: IntoHeaderName>(&self, key: K) -> Option<&HeaderValue>

Gets a value for a given key. If key is not found, Option::None is returned.

Examples

let mut headers = HeaderMap::new();
headers.append("Key", "Value");
let values = headers.get("Key").unwrap();

pub fn get_last<K: IntoHeaderName>(&self, key: K) -> Option<&HeaderValue>

Gets a last value for a given key. If key is not found, Option::None is returned.

Examples

let mut headers = HeaderMap::new();
headers.append("Key", "Value");
let values = headers.get_last("Key").unwrap();

pub fn get_all<K: IntoHeaderName>(&self, key: K) -> GetAll<'_, HeaderValue>

Gets an iterator to the values for a given key.

Examples

let mut headers = HeaderMap::new();
headers.append("Key", "Value1");
headers.append("Key", "Value2");
let mut values = headers.get_all("Key");
let value1 = values.next();
let value2 = values.next();