async_nats::jetstream::consumer

Module pull

Structs

Batch

BatchBuilder

Used for building configuration for a Batch. Created by a Consumer::batch on a Consumer.

BatchConfig

Used for next Pull Request for Pull Consumer

Config

FetchBuilder

Used for building configuration for a Batch with fetch() semantics. Created by a FetchBuilder on a Consumer.

Ordered

OrderedConfig

Configuration for consumers. From a high level, the durable_name and deliver_subject fields have a particularly strong influence on the consumer’s overall behavior.

Sequence

Stream

StreamBuilder

Used for building configuration for a Stream. Created by a Consumer::stream on a Consumer.

Enums

BatchErrorKind

BatchRequestErrorKind

ConsumerRecreateErrorKind

MessagesErrorKind

OrderedErrorKind

Type Aliases

BatchError

BatchRequestError

ConsumerRecreateError

MessagesError

OrderedError