async_nats::jetstream

Module kv

A Key-Value store built on top of JetStream, allowing you to store and retrieve data using simple key-value pairs.

Modules

bucket

Structs

Config

Configuration values for key value stores.

Entry

An entry in a key-value bucket.

History

A structure representing the history of a key-value bucket, yielding past values.

Keys

ParseOperationError

Store

A struct used as a handle for the bucket.

Watch

A structure representing a watch on a key-value bucket, yielding values whenever there are changes.

Enums

CreateErrorKind

EntryErrorKind

Operation

Describes what kind of operation and entry represents

PutErrorKind

StatusErrorKind

UpdateErrorKind

WatchErrorKind

WatcherErrorKind

Type Aliases

CreateError

DeleteError

DeleteErrorKind

EntryError

HistoryError

HistoryErrorKind

PurgeError

PurgeErrorKind

PutError

StatusError

UpdateError

WatchError

WatcherError