async_nats::jetstream

Module object_store

Object Store module

Structs

Config

Configuration values for object store buckets.

List

Object

Represents an object stored in a bucket.

ObjectInfo

Meta and instance information about an object.

ObjectLink

A link to another object, potentially in another bucket.

ObjectMetadata

Meta information about an object.

ObjectOptions

ObjectStore

A blob store capable of storing large objects efficiently in streams.

UpdateMetadata

Watch

Enums

AddLinkErrorKind

DeleteErrorKind

GetErrorKind

InfoErrorKind

PutErrorKind

SealErrorKind

UpdateMetadataErrorKind

WatchErrorKind

WatcherErrorKind

Traits

AsObjectInfo

Type Aliases

AddLinkError

DeleteError

GetError

InfoError

ListError

ListErrorKind

ListerError

ListerErrorKind

PutError

SealError

UpdateMetadataError

WatchError

WatcherError