async_nats::jetstream

Module stream

Manage operations on a Stream, create/delete/update Consumer.

Structs

ClusterInfo

Information about the stream’s, consumer’s associated JetStream cluster

Config

StreamConfig determines the properties for a stream. There are sensible defaults for most. If no subjects are given the name will be used as the only subject.

ConsumerLimits

ConsumerNames

Consumers

DeleteStatus

DirectGetBuilder

DirectGetRequest

External

Info

Shows config and current state for this stream.

InfoWithSubjects

No

PagedInfo

PauseResponse

PeerInfo

The members of the RAFT cluster

Placement

Placement describes on which cluster or tags the stream should be placed.

Purge

PurgeRequest

The payload used to generate a purge request.

PurgeResponse

The response generated by trying to purge a stream.

RawMessage

A raw stream message in the representation it is stored.

RawMessageBuilder

RawMessageRequest

Republish

Source

SourceInfo

State

information about the given stream.

Stream

Handle to operations that can be performed on a Stream. It’s generic over the type of info field to allow Stream with or without info contents.

StreamInfoBuilder

StreamInfoRequest

StreamValue

SubjectTransform

WithHeaders

Marker type indicating that headers should be included in the response.

WithoutHeaders

Marker type indicating that headers should be excluded from the response.

Yes

Enums

Compression

ConsumerCreateStrictErrorKind

ConsumerErrorKind

ConsumerUpdateErrorKind

DeleteMessageErrorKind

DirectGetErrorKind

DiscardPolicy

DiscardPolicy determines how we proceed when limits of messages or bytes are hit. The default, Old will remove older messages. New will fail to store the new message.

LastRawMessageErrorKind

PersistenceMode

Determines the persistence mode for stream data.

PurgeErrorKind

RetentionPolicy

RetentionPolicy determines how messages in a set are retained.

StorageType

determines how messages are stored for retention.

Traits

ToAssign

Type Aliases

ConsumerCreateStrictError

ConsumerError

ConsumerUpdateError

ConsumersError

ConsumersErrorKind

DeleteMessageError

DirectGetError

InfoError

LastRawMessageError

PurgeError

RawMessageError

RawMessageErrorKind