async_nats::jetstream

Struct ErrorCode

pub struct ErrorCode(pub u64);

Tuple Fields
0: u64
Implementations

impl ErrorCode

pub const CLUSTER_PEER_NOT_MEMBER: ErrorCode

Peer not a member

pub const CONSUMER_EPHEMERAL_WITH_DURABLE: ErrorCode

Consumer expected to be ephemeral but detected a durable name set in subject

pub const STREAM_EXTERNAL_DELETE_PREFIX_OVERLAPS: ErrorCode

Stream external delivery prefix overlaps with stream subject

pub const ACCOUNT_RESOURCES_EXCEEDED: ErrorCode

Resource limits exceeded for account

pub const CLUSTER_NOT_AVAILABLE: ErrorCode

Jetstream system temporarily unavailable

pub const STREAM_SUBJECT_OVERLAP: ErrorCode

Subjects overlap with an existing stream

pub const STREAM_WRONG_LAST_SEQUENCE: ErrorCode

Wrong last sequence

pub const NAME_NOT_MATCH_SUBJECT: ErrorCode

Template name in subject does not match request

pub const CLUSTER_NO_PEERS: ErrorCode

No suitable peers for placement

pub const CONSUMER_EPHEMERAL_WITH_DURABLE_NAME: ErrorCode

Consumer expected to be ephemeral but a durable name was set in request

pub const INSUFFICIENT_RESOURCES: ErrorCode

Insufficient resources

pub const MIRROR_MAX_MESSAGE_SIZE_TOO_BIG: ErrorCode

Stream mirror must have max message size >= source

pub const STREAM_DELETE_FAILED: ErrorCode

Generic error from stream deletion operation

pub const BAD_REQUEST: ErrorCode

Bad request

pub const NOT_SUPPORTED_IN_CLUSTER_MODE: ErrorCode

Not currently supported in clustered mode

pub const CONSUMER_NOT_FOUND: ErrorCode

Consumer not found

pub const SOURCE_MAX_MESSAGE_SIZE_TOO_BIG: ErrorCode

Stream source must have max message size >= target

pub const STREAM_ASSIGNMENT: ErrorCode

Generic error when stream operation fails.

pub const STREAM_MESSAGE_EXCEEDS_MAXIMUM: ErrorCode

Message size exceeds maximum allowed

pub const STREAM_CREATE_TEMPLATE: ErrorCode

Generic error for stream creation error with a string

pub const INVALID_JSON: ErrorCode

Invalid JSON

pub const STREAM_INVALID_EXTERNAL_DELIVERY_SUBJECT: ErrorCode

Stream external delivery prefix must not contain wildcards

pub const STREAM_RESTORE: ErrorCode

Restore failed

pub const CLUSTER_INCOMPLETE: ErrorCode

Incomplete results

pub const NO_ACCOUNT: ErrorCode

Account not found

pub const RAFT_GENERAL: ErrorCode

General RAFT error

pub const RESTORE_SUBSCRIBE_FAILED: ErrorCode

Jetstream unable to subscribe to restore snapshot

pub const STREAM_DELETE: ErrorCode

Stream deletion failed

pub const STREAM_EXTERNAL_API_OVERLAP: ErrorCode

Stream external api prefix must not overlap

pub const MIRROR_WITH_SUBJECTS: ErrorCode

Stream mirrors can not contain subjects

pub const JETSTREAM_NOT_ENABLED: ErrorCode

Jetstream not enabled

pub const JETSTREAM_NOT_ENABLED_FOR_ACCOUNT: ErrorCode

Jetstream not enabled for account

pub const SEQUENCE_NOT_FOUND: ErrorCode

Sequence not found

pub const STREAM_MIRROR_NOT_UPDATABLE: ErrorCode

Stream mirror configuration can not be updated

pub const STREAM_SEQUENCE_NOT_MATCH: ErrorCode

Expected stream sequence does not match

pub const STREAM_WRONG_LAST_MESSAGE_ID: ErrorCode

Wrong last msg id

pub const TEMP_STORAGE_FAILED: ErrorCode

Jetstream unable to open temp storage for restore

pub const STORAGE_RESOURCES_EXCEEDED: ErrorCode

Insufficient storage resources available

pub const STREAM_MISMATCH: ErrorCode

Stream name in subject does not match request

pub const STREAM_NOT_MATCH: ErrorCode

Expected stream does not match

pub const MIRROR_CONSUMER_SETUP_FAILED: ErrorCode

Setting up consumer mirror failed

pub const NOT_EMPTY_REQUEST: ErrorCode

Expected an empty request payload

pub const STREAM_NAME_EXIST: ErrorCode

Stream name already in use with a different configuration

pub const CLUSTER_TAGS: ErrorCode

Tags placement not supported for operation

pub const MAXIMUM_CONSUMERS_LIMIT: ErrorCode

Maximum consumers limit reached

pub const SOURCE_CONSUMER_SETUP_FAILED: ErrorCode

General source consumer setup failure

pub const CONSUMER_CREATE: ErrorCode

Consumer creation failed

pub const CONSUMER_DURABLE_NAME_NOT_IN_SUBJECT: ErrorCode

Consumer expected to be durable but no durable name set in subject

pub const STREAM_LIMITS: ErrorCode

Stream limits error

pub const STREAM_REPLICAS_NOT_UPDATABLE: ErrorCode

Replicas configuration can not be updated

pub const STREAM_TEMPLATE_NOT_FOUND: ErrorCode

Template not found

pub const CLUSTER_NOT_ASSIGNED: ErrorCode

Jetstream cluster not assigned to this server

pub const CLUSTER_NOT_LEADER: ErrorCode

Jetstream cluster can’t handle request

pub const CONSUMER_NAME_EXIST: ErrorCode

Consumer name already in use

pub const MIRROR_WITH_SOURCES: ErrorCode

Stream mirrors can’t also contain other sources

pub const STREAM_NOT_FOUND: ErrorCode

Stream not found

pub const CLUSTER_REQUIRED: ErrorCode

Jetstream clustering support required

pub const CONSUMER_DURABLE_NAME_NOT_SET: ErrorCode

Consumer expected to be durable but a durable name was not set

pub const MAXIMUM_STREAMS_LIMIT: ErrorCode

Maximum number of streams reached

pub const MIRROR_WITH_START_SEQUENCE_AND_TIME: ErrorCode

Stream mirrors can not have both start seq and start time configured

pub const STREAM_SNAPSHOT: ErrorCode

Stream snapshot failed

pub const STREAM_UPDATE: ErrorCode

Stream update failed

pub const CLUSTER_NOT_ACTIVE: ErrorCode

Jetstream not in clustered mode

pub const CONSUMER_DURABLE_NAME_NOT_MATCH_SUBJECT: ErrorCode

Consumer name in subject does not match durable name in request

pub const MEMORY_RESOURCES_EXCEEDED: ErrorCode

Insufficient memory resources available

pub const MIRROR_WITH_SUBJECT_FILTERS: ErrorCode

Stream mirrors can not contain filtered subjects

pub const STREAM_CREATE: ErrorCode

Stream create failed with a string

pub const CLUSTER_SERVER_NOT_MEMBER: ErrorCode

Server is not a member of the cluster

pub const NO_MESSAGE_FOUND: ErrorCode

No message found

pub const SNAPSHOT_DELIVER_SUBJECT_INVALID: ErrorCode

Deliver subject not valid

pub const STREAM_GENERALOR: ErrorCode

General stream failure

pub const STREAM_INVALID_CONFIG: ErrorCode

Invalid stream config

pub const STREAM_REPLICAS_NOT_SUPPORTED: ErrorCode

Replicas > 1 not supported in non-clustered mode

pub const STREAM_MESSAGE_DELETE_FAILED: ErrorCode

Stream message delete failed

pub const PEER_REMAP: ErrorCode

Peer remap failed

pub const STREAM_STORE_FAILED: ErrorCode

Stream store failed

pub const CONSUMER_CONFIG_REQUIRED: ErrorCode

Consumer config required

pub const CONSUMER_DELIVER_TO_WILDCARDS: ErrorCode

Consumer deliver subject has wildcards

pub const CONSUMER_PUSH_MAX_WAITING: ErrorCode

Consumer in push mode can not set max waiting

pub const CONSUMER_DELIVER_CYCLE: ErrorCode

Consumer deliver subject forms a cycle

pub const CONSUMER_MAX_PENDING_ACK_POLICY_REQUIRED: ErrorCode

Consumer requires ack policy for max ack pending

pub const CONSUMER_SMALL_HEARTBEAT: ErrorCode

Consumer idle heartbeat needs to be >= 100ms

pub const CONSUMER_PULL_REQUIRES_ACK: ErrorCode

Consumer in pull mode requires ack policy

pub const CONSUMER_PULL_NOT_DURABLE: ErrorCode

Consumer in pull mode requires a durable name

pub const CONSUMER_PULL_WITH_RATE_LIMIT: ErrorCode

Consumer in pull mode can not have rate limit set

pub const CONSUMER_MAX_WAITING_NEGATIVE: ErrorCode

Consumer max waiting needs to be positive

pub const CONSUMER_HEARTBEAT_REQUIRES_PUSH: ErrorCode

Consumer idle heartbeat requires a push based consumer

pub const CONSUMER_FLOW_CONTROL_REQUIRES_PUSH: ErrorCode

Consumer flow control requires a push based consumer

pub const CONSUMER_DIRECT_REQUIRES_PUSH: ErrorCode

Consumer direct requires a push based consumer

pub const CONSUMER_DIRECT_REQUIRES_EPHEMERAL: ErrorCode

Consumer direct requires an ephemeral consumer

pub const CONSUMER_ON_MAPPED: ErrorCode

Consumer direct on a mapped consumer

pub const CONSUMER_FILTER_NOT_SUBSET: ErrorCode

Consumer filter subject is not a valid subset of the interest subjects

pub const CONSUMER_INVALID_POLICY: ErrorCode

Invalid consumer policy

pub const CONSUMER_INVALID_SAMPLING: ErrorCode

Failed to parse consumer sampling configuration

pub const STREAM_INVALID: ErrorCode

Stream not valid

pub const CONSUMER_WQ_REQUIRES_EXPLICIT_ACK: ErrorCode

Workqueue stream requires explicit ack

pub const CONSUMER_WQ_MULTIPLE_UNFILTERED: ErrorCode

Multiple non-filtered consumers not allowed on workqueue stream

pub const CONSUMER_WQ_CONSUMER_NOT_UNIQUE: ErrorCode

Filtered consumer not unique on workqueue stream

pub const CONSUMER_WQ_CONSUMER_NOT_DELIVER_ALL: ErrorCode

Consumer must be deliver all on workqueue stream

pub const CONSUMER_NAME_TOO_LONG: ErrorCode

Consumer name is too long

pub const CONSUMER_BAD_DURABLE_NAME: ErrorCode

Durable name can not contain token separators and wildcards

pub const CONSUMER_STORE_FAILED: ErrorCode

Error creating store for consumer

pub const CONSUMER_EXISTING_ACTIVE: ErrorCode

Consumer already exists and is still active

pub const CONSUMER_REPLACEMENT_WITH_DIFFERENT_NAME: ErrorCode

Consumer replacement durable config not the same

pub const CONSUMER_DESCRIPTION_TOO_LONG: ErrorCode

Consumer description is too long

pub const STREAM_HEADER_EXCEEDS_MAXIMUM: ErrorCode

Header size exceeds maximum allowed of 64k

pub const CONSUMER_WITH_FLOW_CONTROL_NEEDS_HEARTBEATS: ErrorCode

Consumer with flow control also needs heartbeats

pub const STREAM_SEALED: ErrorCode

Invalid operation on sealed stream

pub const STREAM_PURGE_FAILED: ErrorCode

Stream purge failed

pub const STREAM_ROLLUP_FAILED: ErrorCode

Stream rollup failed

pub const CONSUMER_INVALID_DELIVER_SUBJECT: ErrorCode

Invalid push consumer deliver subject

pub const STREAM_MAX_BYTES_REQUIRED: ErrorCode

Account requires a stream config to have max bytes set

pub const CONSUMER_MAX_REQUEST_BATCH_NEGATIVE: ErrorCode

Consumer max request batch needs to be > 0

pub const CONSUMER_MAX_REQUEST_EXPIRES_TO_SMALL: ErrorCode

Consumer max request expires needs to be >= 1ms

pub const CONSUMER_MAX_DELIVER_BACKOFF: ErrorCode

Max deliver is required to be > length of backoff values

pub const STREAM_INFO_MAX_SUBJECTS: ErrorCode

Subject details would exceed maximum allowed

pub const STREAM_OFFLINE: ErrorCode

Stream is offline

pub const CONSUMER_OFFLINE: ErrorCode

Consumer is offline

pub const NO_LIMITS: ErrorCode

No jetstream default or applicable tiered limit present

pub const CONSUMER_MAX_PENDING_ACK_EXCESS: ErrorCode

Consumer max ack pending exceeds system limit

pub const STREAM_MAX_STREAM_BYTES_EXCEEDED: ErrorCode

Stream max bytes exceeds account limit max stream bytes

pub const STREAM_MOVE_AND_SCALE: ErrorCode

Can not move and scale a stream in a single update

pub const STREAM_MOVE_IN_PROGRESS: ErrorCode

Stream move already in progress

pub const CONSUMER_MAX_REQUEST_BATCH_EXCEEDED: ErrorCode

Consumer max request batch exceeds server limit

pub const CONSUMER_REPLICAS_EXCEEDS_STREAM: ErrorCode

Consumer config replica count exceeds parent stream

pub const CONSUMER_NAME_CONTAINS_PATH_SEPARATORS: ErrorCode

Consumer name can not contain path separators

pub const STREAM_NAME_CONTAINS_PATH_SEPARATORS: ErrorCode

Stream name can not contain path separators

pub const STREAM_MOVE_NOT_IN_PROGRESS: ErrorCode

Stream move not in progress

pub const STREAM_NAME_EXIST_RESTORE_FAILED: ErrorCode

Stream name already in use, cannot restore

pub const CONSUMER_CREATE_FILTER_SUBJECT_MISMATCH: ErrorCode

Consumer create request did not match filtered subject from create subject

pub const CONSUMER_CREATE_DURABLE_AND_NAME_MISMATCH: ErrorCode

Consumer durable and name have to be equal if both are provided

pub const REPLICAS_COUNT_CANNOT_BE_NEGATIVE: ErrorCode

Replicas count cannot be negative

pub const CONSUMER_REPLICAS_SHOULD_MATCH_STREAM: ErrorCode

Consumer config replicas must match interest retention stream’s replicas

pub const CONSUMER_METADATA_LENGTH: ErrorCode

Consumer metadata exceeds maximum size

pub const CONSUMER_DUPLICATE_FILTER_SUBJECTS: ErrorCode

Consumer cannot have both filter_subject and filter_subjects specified

pub const CONSUMER_MULTIPLE_FILTERS_NOT_ALLOWED: ErrorCode

Consumer with multiple subject filters cannot use subject based api

pub const CONSUMER_OVERLAPPING_SUBJECT_FILTERS: ErrorCode

Consumer subject filters cannot overlap

pub const CONSUMER_EMPTY_FILTER: ErrorCode

Consumer filter in filter_subjects cannot be empty

pub const SOURCE_DUPLICATE_DETECTED: ErrorCode

Duplicate source configuration detected

pub const SOURCE_INVALID_STREAM_NAME: ErrorCode

Sourced stream name is invalid

pub const MIRROR_INVALID_STREAM_NAME: ErrorCode

Mirrored stream name is invalid

pub const MIRROR_WITH_FIRST_SEQ: ErrorCode

Stream mirrors cannot have first sequence configured

pub const SOURCE_MULTIPLE_FILTERS_NOT_ALLOWED: ErrorCode

Source with multiple subject transforms cannot also have a single subject filter

pub const SOURCE_INVALID_SUBJECT_FILTER: ErrorCode

Source subject filter is invalid

pub const SOURCE_INVALID_TRANSFORM_DESTINATION: ErrorCode

Source transform destination is invalid

pub const SOURCE_OVERLAPPING_SUBJECT_FILTERS: ErrorCode

Source filters cannot overlap

pub const CONSUMER_ALREADY_EXISTS: ErrorCode

Consumer already exists

pub const CONSUMER_DOES_NOT_EXIST: ErrorCode

Consumer does not exist

pub const MIRROR_MULTIPLE_FILTERS_NOT_ALLOWED: ErrorCode

Mirror with multiple subject transforms cannot also have a single subject filter

pub const MIRROR_INVALID_SUBJECT_FILTER: ErrorCode

Mirror subject filter is invalid

pub const MIRROR_OVERLAPPING_SUBJECT_FILTERS: ErrorCode

Mirror subject filters cannot overlap

pub const CONSUMER_INACTIVE_THRESHOLD_EXCESS: ErrorCode

Consumer inactive threshold exceeds system limit

pub const MIRROR_INVALID_TRANSFORM_DESTINATION: ErrorCode

Mirror invalid transform destination

pub const STREAM_TRANSFORM_INVALID_SOURCE: ErrorCode

Stream transform invalid source

pub const STREAM_TRANSFORM_INVALID_DESTINATION: ErrorCode

Stream transform invalid destination

pub const PEDANTIC: ErrorCode

Pedantic error

pub const STREAM_DUPLICATE_MESSAGE_CONFLICT: ErrorCode

Stream duplicate message conflict

pub const CONSUMER_PRIORITY_POLICY_WITHOUT_GROUP: ErrorCode

Consumer priority policy without group

pub const CONSUMER_INVALID_PRIORITY_GROUP: ErrorCode

Consumer invalid priority group

pub const CONSUMER_EMPTY_GROUP_NAME: ErrorCode

Consumer empty group name

pub const CONSUMER_INVALID_GROUP_NAME: ErrorCode

Consumer invalid group name

pub const STREAM_EXPECTED_LAST_SEQ_PER_SUBJECT_NOT_READY: ErrorCode

Stream expected last seq per subject not ready

pub const STREAM_WRONG_LAST_SEQUENCE_CONSTANT: ErrorCode

Stream wrong last sequence constant

pub const MESSAGE_TTL_INVALID: ErrorCode

Message TTL invalid

pub const MESSAGE_TTL_DISABLED: ErrorCode

Message TTL disabled

pub const STREAM_TOO_MANY_REQUESTS: ErrorCode

Stream too many requests

pub const MESSAGE_INCR_DISABLED: ErrorCode

Message increment disabled

pub const MESSAGE_INCR_MISSING: ErrorCode

Message increment missing

pub const MESSAGE_INCR_PAYLOAD: ErrorCode

Message increment payload error

pub const MESSAGE_INCR_INVALID: ErrorCode

Message increment invalid

pub const MESSAGE_COUNTER_BROKEN: ErrorCode

Message counter broken

pub const MIRROR_WITH_COUNTERS: ErrorCode

Mirror with counters

pub const ATOMIC_PUBLISH_DISABLED: ErrorCode

Atomic publish disabled

pub const ATOMIC_PUBLISH_MISSING_SEQ: ErrorCode

Atomic publish sequence missing

pub const ATOMIC_PUBLISH_INCOMPLETE_BATCH: ErrorCode

Atomic publish batch incomplete

pub const ATOMIC_PUBLISH_UNSUPPORTED_HEADER: ErrorCode

Atomic publish unsupported header

pub const CONSUMER_PUSH_WITH_PRIORITY_GROUP: ErrorCode

Consumer push with priority group

pub const ATOMIC_PUBLISH_INVALID_BATCH_ID: ErrorCode

Atomic publish batch ID invalid

pub const STREAM_MIN_LAST_SEQ: ErrorCode

Stream min last sequence error

pub const CONSUMER_ACK_POLICY_INVALID: ErrorCode

Consumer ack policy invalid

pub const CONSUMER_REPLAY_POLICY_INVALID: ErrorCode

Consumer replay policy invalid

pub const CONSUMER_ACK_WAIT_NEGATIVE: ErrorCode

Consumer ack wait negative

pub const CONSUMER_BACKOFF_NEGATIVE: ErrorCode

Consumer backoff negative

pub const REQUIRED_API_LEVEL: ErrorCode

Required API level

pub const MIRROR_WITH_MSG_SCHEDULES: ErrorCode

Mirror with message schedules

pub const SOURCE_WITH_MSG_SCHEDULES: ErrorCode

Source with message schedules

pub const MESSAGE_SCHEDULES_DISABLED: ErrorCode

Message schedules disabled

pub const MESSAGE_SCHEDULES_PATTERN_INVALID: ErrorCode

Message schedules pattern invalid

pub const MESSAGE_SCHEDULES_TARGET_INVALID: ErrorCode

Message schedules target invalid

pub const MESSAGE_SCHEDULES_TTL_INVALID: ErrorCode

Message schedules TTL invalid

pub const MESSAGE_SCHEDULES_ROLLUP_INVALID: ErrorCode

Message schedules rollup invalid

pub const STREAM_EXPECTED_LAST_SEQ_PER_SUBJECT_INVALID: ErrorCode

Stream expected last seq per subject invalid

pub const STREAM_OFFLINE_REASON: ErrorCode

Stream offline reason

pub const CONSUMER_OFFLINE_REASON: ErrorCode

Consumer offline reason

pub const CONSUMER_PRIORITY_GROUP_WITH_POLICY_NONE: ErrorCode

Consumer priority group with policy none

pub const CONSUMER_PINNED_TTL_WITHOUT_PRIORITY_POLICY_NONE: ErrorCode

Consumer pinned TTL without priority policy none

pub const MIRROR_WITH_ATOMIC_PUBLISH: ErrorCode

Mirror with atomic publish

pub const ATOMIC_PUBLISH_TOO_LARGE_BATCH: ErrorCode

Atomic publish batch too large

pub const ATOMIC_PUBLISH_INVALID_BATCH_COMMIT: ErrorCode

Atomic publish batch commit invalid