async_nats

Struct ServerAddr

pub struct ServerAddr(/* private fields */);

Address of a NATS server.

Implementations

impl ServerAddr

pub fn from_url(url: Url) -> Result<Self>

Check if the URL is a valid NATS server address.

pub fn into_inner(self) -> Url

Turn the server address into a standard URL.

pub fn tls_required(&self) -> bool

Returns if tls is required by the client for this server.

pub fn has_user_pass(&self) -> bool

Returns if the server url had embedded username and password.

pub fn scheme(&self) -> &str

pub fn host(&self) -> &str

Returns the host.

pub fn is_websocket(&self) -> bool

pub fn port(&self) -> u16

Returns the port. Delegates to Url::port_or_known_default and defaults to 4222 if none was explicitly specified in creating this ServerAddr.

pub fn as_url_str(&self) -> &str

Returns the URL string.

pub fn username(&self) -> Option<&str>

Returns the optional username in the url.

pub fn password(&self) -> Option<&str>

Returns the optional password in the url.

pub async fn socket_addrs( &self, ) -> Result<impl Iterator<Item = SocketAddr> + '_>

Return the sockets from resolving the server address.