async_nats

Struct ServerInfo

pub struct ServerInfo {

    pub server_id: String,
    pub server_name: String,
    pub host: String,
    pub port: u16,
    pub version: String,
    pub auth_required: bool,
    pub tls_required: bool,
    pub max_payload: usize,
    pub proto: i8,
    pub client_id: u64,
    pub go: String,
    pub nonce: String,
    pub connect_urls: Vec<String>,
    pub client_ip: String,
    pub headers: bool,
    pub lame_duck_mode: bool,
    pub cluster: Option<String>,
    pub domain: Option<String>,
    pub jetstream: bool,

}

Information sent by the server back to this client during initial connection, and possibly again later.

Fields
server_id: String

The unique identifier of the NATS server.

server_name: String

Generated Server Name.

host: String

The host specified in the cluster parameter/options.

port: u16

The port number specified in the cluster parameter/options.

version: String

The version of the NATS server.

auth_required: bool

If this is set, then the server should try to authenticate upon connect.

tls_required: bool

If this is set, then the server must authenticate using TLS.

max_payload: usize

Maximum payload size that the server will accept.

proto: i8

The protocol version in use.

client_id: u64

The server-assigned client ID. This may change during reconnection.

go: String

The version of golang the NATS server was built with.

nonce: String

The nonce used for nkeys.

connect_urls: Vec<String>

A list of server urls that a client can connect to.

client_ip: String

The client IP as known by the server.

headers: bool

Whether the server supports headers.

lame_duck_mode: bool

Whether server goes into lame duck mode.

cluster: Option<String>

Name of the cluster if the server is in cluster-mode

domain: Option<String>

The configured NATS domain of the server.

jetstream: bool

Whether the server supports JetStream.