async_nats

Trait ToServerAddrs

pub trait ToServerAddrs {
    type Iter: Iterator<Item = ServerAddr>;

    // Required method
    fn to_server_addrs(&self) -> Result<Self::Iter>;
}

Capability to convert into a list of NATS server addresses.

There are several implementations ensuring the easy passing of one or more server addresses to functions like crate::connect().

Required Associated Types

type Iter: Iterator<Item = ServerAddr>

Returned iterator over socket addresses which this type may correspond to.

Required Methods

fn to_server_addrs(&self) -> Result<Self::Iter>